refactor(web): migrate to FSD
