+++ apps/web/src/routes/_public.tsx
@@ -1,6 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-
-import PublicLayout from "../layouts/public-layout";
+import PublicLayout from "#/shared/ui/layouts/public-layout.tsx";
 
 export const Route = createFileRoute("/_public")({
     component: PublicLayout,
