+++ apps/web/src/widgets/footer/ui/footer.module.css
