+++ apps/web/src/widgets/footer/ui/footer.tsx
