+++ apps/web/src/widgets/footer/ui/index.ts
